v3.3.2 release -- moved libraries, fixed small crashes, updates to script API
Former-commit-id: babc48bd7a6ab51e23ce271b5bf47c855ecedfc6

+++ Ana/Source/ScriptEngine/ScriptManager.cs
@@ -100,7 +100,7 @@
                         elapsedTime = currentTime - previousTime;
 
                         // Call the update function, giving the elapsed milliseconds since the previous call
-                        ScriptObject.OnUpdate(elapsedTime.TotalMilliseconds);
+                        ScriptObject.OnUpdate((Single)elapsedTime.TotalMilliseconds);
 
                         previousTime = currentTime;
